Use latest version
There are nine packages available in buildpack.toml, but they are all at the same version level. I believe that previously we had tried to support multiple major versions, but this adds some complexity and scope to what we have to maintain given that it's nine packages per version level we want to support. For now, we are only supporting the most recent major version so I'm changing this default version to '*' which will pick up the latest version.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>