Update buildpack.toml
These changes are to allow the buildpack to support multiple version of the liberty runtime. The liberty versions follow the pattern like 22.0.0.1 but that isn't compatible with semver. In the buildpack.toml I want to "translate" 22.0.0.1 to 22.0.1, 22.0.0.2 to 22.0.2, etc.

The variable BP_OPENLIBERTY_VERSION would need to be specified as 22.0.1 or 22.* to get the specific 22.0.1 or latest 22 version, respectfully.

Eventually, we would want to keep the .3, .6, .9, and .12 versions the default for 3 months but add the latest version as an alternate when they come out. Liberty is on a 4-week release cycle. So, for example, if this buildpack were available today we would have the default to be 21.0.12 and the alternate be 22.0.2. When liberty releases 22.0.0.3 we will want to change the default to 22.0.3 for the next 3 4-week cycles and just change the alternate every 4 weeks.