Uses a more specific regex to match CPE & PURL during udpates
Prior to this PR, we are getting spurious PRs that incorrectly change the version in the CPE & PURL. The reason is that the default regex for updating the CPE & PURL is the version regex, which matches three dot separated digits. The version numbers in the CPE & PURL are four dot separate digits, so the default regex matches the first three digits and updates just that part, which results in a bad PR.

This switches the CPE & PURL to using four dot separate digits as the version regex. This should match the full version number and cause it to update correctly.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>