Check if build runs earlier
The buildpack is printing to STDOUT prior to making the determination that it should actually run. This can be confusing to users as output is print, making it look like the buildpack is running, but then the buildpack doesn't actually run. This change moves first output to after the buildpack decides if it will run. This means when the buildpack doesn't run, build is silent (except when debug is enabled).

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>