Move JRE to launch, add JDK for build
This change moves the JRE to be a launch-only layer. This addresses issues when integrating Liberty with the larger Java composite buildpack. Because we need Java during build, this change also adds the JDK as a build + cache layer, which will make it available during build.

This will help because it reduces the caching requirements of the JRE layer across all builds and all buildpacks, not just Liberty. It comes at the expense of requiring the JDK to be downloaded when building a WAR file from pre-compiled bits (does not impact building from source).

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>